requirements.txt
boto3
python-dotenv
PyMuPDF
pytesseract
pillow